+++ plugins/supabase/iac/index.ts
@@ -705,7 +705,7 @@ export const runInit = async ({ build }: { build: BuildType }) => {
 
   p.log.message(
     `Next step: ${link(
-      "https://hot-updater.dev/docs/managed/supabase#step-4-add-hotupdater-to-your-project",
+      "https://docs.hot-updater.dev/docs/managed/supabase#step-4-add-hotupdater-to-your-project",
     )}`,
   );
   p.log.success("Done! 🎉");