+++ plugins/plugin-core/CHANGELOG.md
@@ -553,7 +553,7 @@
 
 ### Patch Changes
 
-- 2b408f2: docs: revamp hot-updater.dev
+- 2b408f2: docs: revamp docs.hot-updater.dev
   - @hot-updater/core@0.21.7
 
 ## 0.21.6